WIP fix ageffod right update

+++ scripts/update_rights.php
@@ -50,45 +50,68 @@ function fixAgefoddRights($TRights = array(), $mod_number = 0)
 					."'".$right->type."',"
 					.$right->bydefault." );";
 
-			}
-
-			$res = $db->query($sqlrd);
-			if (!$res) {
-				$fail = true;
-				break 2;
+				$res = $db->query($sqlrd);
+				if (!$res) {
+					$fail = true;
+					break 2;
+				}
 			}
 
 			// UPDATE des droits user
-			$sqlur = "UPDATE ".MAIN_DB_PREFIX."user_rights SET fk_id=".$mod_number.(intval(substr($r_id, -2)) -1)." WHERE fk_id = ".$r_id." AND entity = ".$entity.";";
-			$res = $db->query($sqlur);
-			if (!$res) {
-				$fail = true;
-				break 2;
+			// 1. check no Duplicate entry for key 'uk_user_rights'
+			$res = $db->query("SELECT COUNT(*) as countNumb ".MAIN_DB_PREFIX."user_rights WHERE fk_id=".$mod_number.(intval(substr($r_id, -2)) -1)." AND entity = ".$entity.";");
+			if ($res) {
+				$count = $db->fetch_object($res);
+				if(empty($count->countNumb))
+				{
+					// 2. UPDATE des droits user
+					$sqlur = "UPDATE ".MAIN_DB_PREFIX."user_rights SET fk_id=".$mod_number.(intval(substr($r_id, -2)) -1)." WHERE fk_id = ".$r_id." AND entity = ".$entity.";";
+					$res = $db->query($sqlur);
+					if (!$res) {
+						$fail = true;
+						break 2;
+					}
+				}
 			}
 
-			// UPDATE des droits groupes
-			$sqlug = "UPDATE ".MAIN_DB_PREFIX."usergroup_rights SET fk_id=".$mod_number.(intval(substr($r_id, -2)) -1)." WHERE fk_id = ".$r_id." AND entity = ".$entity.";";
-			$res = $db->query($sqlug);
-			if (!$res) {
-				$fail = true;
-				break 2;
+			// UPDATE des droits group
+			// 1. check no Duplicate entry for key 'uk_group_rights'
+			$res = $db->query("SELECT COUNT(*) as countNumb ".MAIN_DB_PREFIX."usergroup_rights WHERE fk_id=".$mod_number.(intval(substr($r_id, -2)) -1)." AND entity = ".$entity.";");
+			if ($res) {
+				$count = $db->fetch_object($res);
+				if(empty($count->countNumb))
+				{
+					// 2. UPDATE des droits groupes
+					$sqlug = "UPDATE ".MAIN_DB_PREFIX."usergroup_rights SET fk_id=".$mod_number.(intval(substr($r_id, -2)) -1)." WHERE fk_id = ".$r_id." AND entity = ".$entity.";";
+					$res = $db->query($sqlug);
+					if (!$res) {
+						$fail = true;
+						break 2;
+					}
+				}
 			}
+
+
 
 		}
 		$TRights_id[] = $r_id;
 
 	}
 
-	// suppression des droits erronés
-	$sql = "DELETE FROM ".MAIN_DB_PREFIX."rights_def WHERE id IN (".implode(', ', $TRights_id).");";
+	if(!empty($TRights_id) and is_array($TRights_id))
+	{
+		// suppression des droits erronés
+		$sql = "DELETE FROM ".MAIN_DB_PREFIX."rights_def WHERE id IN (".implode(', ', $TRights_id).");";
 
-	$resql = $db->query($sql);
-	if (!$resql) $fail = true;
+		$resql = $db->query($sql);
+		if (!$resql) $fail = true;
+	}
 
 	if ($fail)
 	{
-//		print_r($db->lastqueryerror);
-//		var_dump($db->lasterror()); exit;
+		print_r($db->lastqueryerror);
+		print_r($db->lastquery);
+		var_dump($db->lasterror()); exit;
 		$db->rollback();
 		return -2;
 	}
